Move constants to setup
