+++ corebehrt/functional/outcomes.py
@@ -1,4 +1,4 @@
-from corebehrt.common.constants import PID_COL, ABSPOS_COL
+from corebehrt.modules.setup.constants import PID_COL, ABSPOS_COL
 import pandas as pd
